code split resource controller and definition
